% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.posim.R
\name{animate,posim-method}
\alias{animate,posim-method}
\title{Animates simulated movement tracks}
\usage{
\S4method{animate}{posim}(
  x,
  select = NULL,
  steps = NULL,
  tail.length = 10,
  blink.duration = 12,
  delay = 1/24,
  filename = "track.gif"
)
}
\arguments{
\item{x}{The returned object from a call to \code{\link[=posim]{posim()}}}

\item{select}{Vector with indices denoting the ids/indices of which individuals to animate. Defaults to NULL, which animates all individuals.}

\item{steps}{The number of steps to animate.}

\item{tail.length}{}
}
\description{
Creates an animated GIF file showing in each frame only the most recent movement tracks
}
