% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClasses.R
\docType{class}
\name{posim-class}
\alias{posim-class}
\title{S4 class to represent IBM simulation results}
\description{
S4 class to represent IBM simulation results
}
\section{Slots}{

\describe{
\item{\code{conf}}{List specifying simulation options}

\item{\code{filename}}{Character specifying filename of serialized ascii data (sasc)}

\item{\code{data}}{Dataframe containing data from sasc file}

\item{\code{tracks}}{Dataframe containing movement tracks for any followed porpoises in simulation}

\item{\code{raster}}{A RasterLayer object based on bathymetric data}
}}

