\name{posim}
\alias{posim}
\alias{posim}
\docType{package}
\title{
  Individual based harbour porpoise model
}
\description{
  Provides a framework for running a harbour porpoise IBM
  with customized data.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
André Moan, <andre.moan@hi.no>.

Maintainer: André Moan <andre.moan@hi.no>
}
\references{
  None.
}
\keyword{ package }
\seealso{
  TODO: add links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
